Inspections (#141)
* Add serial version to serialisable classes.  Then don't use it properly.

* Make final-able things final

* Make final-able things final

* Mis sorted modifiers

* Redundant field initialisation

* Redundant constructor

* size()==0 becomes isEmpty()

* Nullsafe comparison ordering

* ? extends Object == Object

* Remove unnecessary super()

* Redundant .toString()

* Unnecessarily qualified static import

* Inverted if statement

* Simplify negated equality

* Simplify out if statements

* Make final

* Inline one use variables

* Final

* Default annotations

* Ensure exception root causes are passed on

* Formally ignore some ignored exceptions

* Unused Import

* Remove redundant abstract method

* Identical to super method it overrides

* Public abstract constructors become protected

* Javadoc line breaks

* Enhanced switch statement

* Use pattern variables

* Use text blocks

* String contains() usage

* Comparator statics

* Use putAll on collection cloning

* Comparator

* Use String.valueOf() rather than +""

* String.equals versus String.isEmpty

* Rename IC/PC

* Variable name fixes

* Inherit variable names hierarchically

* Variable name fixup

* Variable name fixup

* Numeric cleanups

* Early break

* static!

* entrySet iteration

* Use EnumMaps where possible

* indexOf char rather than String where possible

* String Builders

* Code not in use

* Initialisations

* serialisations

* Inspections

* Make final

* Use isEmpty()

* Redundant type casts

* Weak type unnecessary cast

* Remove redundancy

* Regex simplification

+++ src/net/coagulate/GPHUD/Data/DateTime.java
@@ -176,7 +176,7 @@ public abstract class DateTime {
 	// ----- Internal Statics -----
 	@Nonnull
 	private static String fromUnixTime(final int date,
-	                                   @Nonnull final DateFormat df) { return df.format(new Date(((long) (date))*((long) 1000))); }
+	                                   @Nonnull final DateFormat df) { return df.format(new Date((date) *(1000L))); }
 
 	@Nonnull
 	private static TextInput ti(final String prefix,
